@@ -1,7 +1,7 @@
 name: railroad_crossing
 description: railroad crossing simulator
 publish_to: 'none'
-version: 1.1.0+20
+version: 1.2.1+31
 
 environment:
   sdk: '>=2.18.0 <4.0.0'
@@ -10,32 +10,32 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  webview_flutter_android: ^3.15.0
+  webview_flutter_android: ^3.16.9
   webview_flutter: ^4.0.0
   # Firebase Analytics
-  firebase_core: ^3.6.0
-  firebase_analytics: ^11.3.3
-  firebase_app_check: ^0.3.1+4
+  firebase_core: ^3.9.0
+  firebase_analytics: ^11.3.6
+  firebase_app_check: ^0.3.1+7
   # App Icon
   cupertino_icons: ^1.0.8
-  flutter_launcher_icons: ^0.14.1
+  flutter_launcher_icons: ^0.14.2
   # Audio Player
   audioplayers: ^6.1.0
   # Vibration
   vibration: ^2.0.1
   # SharedPreference
-  shared_preferences: ^2.3.2
+  shared_preferences: ^2.3.4
   # Floating Button
   fab_circular_menu_plus: ^0.0.2
   # Game Center
-  # games_services: ^4.0.2
+#  games_services: ^4.0.2
   # Admob
   google_mobile_ads: ^5.2.0
-  app_tracking_transparency: ^2.0.6
+  app_tracking_transparency: ^2.0.6+1
   # Device locale
   devicelocale: ^0.8.1
   # OS version
-  device_info_plus: ^11.1.0
+  device_info_plus: ^11.2.0
   # Localize
   flutter_localizations:
     sdk: flutter
@@ -43,10 +43,10 @@ dependencies:
   #env
   flutter_dotenv: ^5.2.1
   # Revenue Cat
-  purchases_flutter: ^8.2.2
+  purchases_flutter: ^8.4.0
   url_launcher: ^6.3.1
   #MVVM
-  hooks_riverpod: ^2.5.3
+  hooks_riverpod: ^2.6.1
   flutter_hooks: ^0.20.5
   #Generative AI
   googleapis_auth: ^1.6.0
@@ -55,16 +55,16 @@ dependencies:
   ntp: ^2.0.0
   #Save Image
   permission_handler: ^11.3.1
-  path_provider: ^2.1.4
+  path_provider: ^2.1.5
   image_gallery_saver: ^2.0.3
-  share_plus: ^10.0.3
+  share_plus: ^10.1.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_lints: ^5.0.0
   # App Splash
-  flutter_native_splash: ^2.4.1
+  flutter_native_splash: ^2.4.3
 
 flutter_icons:
   image_path: "assets/images/appIcon.png"
